Fixes >2G address problem and adds test for it. Changes LZ4ByteBufferUtils to be generated from lz4 utils template.
